pyproject.toml
[project]
name = "grafana"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "loki-logger-handler>=1.1.1",
    "nextlog>=2.0.1",
    "python-logging-loki-v2>=0.4.4",
]

[dependency-groups]
dev = [
    "icecream>=2.1.4",
]
